Refactor: Move offer/wanted to existing routes - move doc - move tests - move logic - adjust frontend
