+++ backend/index.js
@@ -18,7 +18,7 @@ const vehicle = require('./routes/vehicle');
 app.use('', vehicle.router);
 
 const offer = require('./routes/offer');
-app.use('', offer.router);
+app.use('/offer', offer.router);
 
 const wanted = require('./routes/wanted');
 app.use('', wanted.router);
