+++ backend/test/offer.test.js
@@ -1,5 +1,5 @@
 const {expect, test, afterAll} = require('@jest/globals');
-const ad = require('../routes/ad');
+const offer = require('../routes/offer');
 const mariadb = require('mariadb');
 
 const pool = mariadb.createPool({
@@ -9,7 +9,8 @@ const pool = mariadb.createPool({
     password: 'admin',
     database: 'cargodb',
 });
-test('get ad by correct Id', async () =>{
+
+test('get offer by correct Id', async () =>{
     let conn;
     try {
         conn = await pool.getConnection();
@@ -21,24 +22,14 @@ test('get ad by correct Id', async () =>{
                 VALUES (10,'Car1', 4, 500.0, 'car1.jpg', '2x2x2', 'GPS, Bluetooth', 10)`);
         await conn.query(`INSERT INTO offer (offerId, vehicleId, adId, pricePerPerson, pricePerFreight) \
                 VALUES (10, 10, 10, 50.0, 100.0)`);
-        const res = (await ad.getAdById(10));
+        const res = (await offer.getOfferById(10));
         expect(res.data).toEqual(
             {
+                'offerId': 10,
+                'vehicleId': 10,
                 'adId': 10,
-                'description': 'Ja Beschreibung halt so lololol',
-                'startLocation': 'City A',
-                'endLocation': 'City B',
-                'type': 'offer',
-                'intermediateGoals': [],
-                'startDate': new Date('2023-01-10'),
-                'endDate': new Date('2023-01-15'),
-                'animals': 0,
-                'smoker': 1,
-                'notes': 'No pets allowed',
-                'numSeats': 4,
-                'picture': null,
-                'active': 1,
-                'userId': 10,
+                'pricePerPerson': 50.0,
+                'pricePerFreight': 100.0,
             });
     } finally {
         conn.query(`DELETE FROM ad WHERE adId = 10`);
@@ -49,8 +40,8 @@ test('get ad by correct Id', async () =>{
     };
 });
 
-test('get ad by wrong Id', async () =>{
-    const res2 = await ad.getAdById(99999999999999999);
+test('get offer by wrong Id', async () =>{
+    const res2 = await offer.getOfferById(99999999999999999);
     expect(res2).toEqual({success: false});
 });
